+++ CHANGELOG.md
@@ -247,3 +247,4 @@
 - 2024-01-11: Impl bettercontact, LICENSE, frontier, find email, +6 query
 - 2024-01-11: Retry configuration, lead, email opener, NOTICE, +6
 - 2024-01-12: Impl descend, mutate, discover, deal, +4 query
+- 2024-01-13: Bench daemon, ready pool, discovery, NOTICE, +2